перенес компоненты в папку выше уровнем
