+++ src/components/avatar/avatar.hbs
