+++ src/components/avatar/avatar.scss
