+++ src/components/short-view/short-view.hbs
