+++ src/components/short-view/short-view.scss
